pyproject.toml
[project]
name = "langchain-course"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.12.0",
    "isort>=7.0.0",
    "langchain>=1.2.0",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.3",
    "langchain-tavily>=0.2.15",
    "langchainhub>=0.1.0",
    "langsmith>=0.1.0",
    "python-dotenv>=1.2.1",
]
